requirements.txt
discord.py
python-dotenv
spotipy
requests
pymongo
beautifulsoup4==4.9.3
aiohttp
asyncio
pytz
matplotlib